+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- includeCompleted flag for v1/characters/{character_id}/industry/jobs
+
 ## [4.2.0] - 2019-09-12
 ### Added
 - Types for blueprint route location_flag property.